tests: Make the test harness self-contained
Remove Boost from the tests and provide the minimal framework and build glue needed to exercise the stripped source tree without Coin's former infrastructure.

+++ src/base/heap.cpp
@@ -380,7 +380,7 @@ public:
 
   static void heap_print_cb(void * v, SbString& str) {
     wrapped_value* value = reinterpret_cast<wrapped_value*>(v);
-    str += std::to_string(value->x).c_str();
+    str += std::to_string(static_cast<int>(value->x)).c_str();
   }
 
   static double heap_evaluate_cb(void * v)
